# docs/swagger.yaml
definitions:
...
  entities.Dish:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      dishId:
        type: integer
      imageUrl:
        type: string
      isAvailable:
        type: boolean
      isVeg:
        type: boolean
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      salePrice:
        type: number
      sellerId:
        type: integer
    type: object
...
  req.UpdateDishReq:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      isAvailable:
        type: boolean
      isVeg:
        type: boolean
      name:
        minLength: 3
        type: string
      price:
        minimum: 0
        type: number
      quantity:
        minimum: 0
        type: integer
      salePrice:
        minimum: 0
        type: number
    required:
    - categoryId
    - name
    - price
    - quantity
    type: object
...
  res.CommonRes:
    properties:
      error:
        type: string
      message:
        type: string
      result: {}
      status:
        type: string
    type: object
  res.DishListRes:
    properties:
      dishList:
        items:
          $ref: '#/definitions/entities.Dish'
        type: array
      message:
        type: string
      status:
        type: string
    type: object
...
  res.SingleDishRes:
    properties:
      dish:
        $ref: '#/definitions/entities.Dish'
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /dishes:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of dishes for the user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: p
        type: string
      - description: Number of items per page
        in: query
        name: l
        type: string
      - description: Item category
        in: query
        name: category
        type: string
      - description: Item by seller
        in: query
        name: seller
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched dishes
          schema:
            $ref: '#/definitions/res.DishListRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Get paginated list of dishes
      tags:
      - Common
  /dishes/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific dish by ID for the user
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched dish
          schema:
            $ref: '#/definitions/res.SingleDishRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Get a dish
      tags:
      - Common
...
  /search/dishes:
    get:
      consumes:
      - application/json
      description: Search for dishes based on a query
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched dishes
          schema:
            $ref: '#/definitions/res.DishListRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      summary: Search dishes
      tags:
      - Common
...
  /seller/addDish:
    post:
      consumes:
      - multipart/form-data
      description: Add a new dish for the seller
      parameters:
      - description: Image file for the dish
        in: formData
        name: image
        required: true
        type: file
      - in: formData
        name: categoryId
        required: true
        type: integer
      - in: formData
        name: description
        type: string
      - in: formData
        name: isAvailable
        type: boolean
      - in: formData
        name: isVeg
        type: boolean
      - in: formData
        minLength: 3
        name: name
        required: true
        type: string
      - in: formData
        minimum: 0
        name: price
        required: true
        type: number
      - in: formData
        minimum: 0
        name: quantity
        required: true
        type: integer
      - in: formData
        name: salePrice
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created dish
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Create a dish
      tags:
      - Seller Dish
  /seller/dishes:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all dishes for the seller
      parameters:
      - description: Category Id
        in: query
        name: category
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched dishes
          schema:
            $ref: '#/definitions/res.DishListRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Get all dishes
      tags:
      - Seller Dish
  /seller/dishes/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a specific dish by ID for the seller
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted dish
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Delete a dish
      tags:
      - Seller Dish
    get:
      consumes:
      - application/json
      description: Retrieve a specific dish by ID for the seller
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched dish
          schema:
            $ref: '#/definitions/res.SingleDishRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Get a dish
      tags:
      - Seller Dish
    put:
      consumes:
      - application/json
      description: Modify an existing dish by ID for the seller
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      - description: Dish update request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/req.UpdateDishReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated dish
          schema:
            $ref: '#/definitions/res.CommonRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.CommonRes'
        "401":
          description: Unauthorized Access
          schema:
            $ref: '#/definitions/res.CommonRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.CommonRes'
      security:
      - Bearer: []
      summary: Update a dish
      tags:
      - Seller Dish
...
securityDefinitions:
  Bearer:
    description: 'Authentication using a JSON Web Token (JWT). The token should be
      included in the request header named "Authorization". The format of the header
      is: Authorization: Bearer <token>. Replace `<token>` with the actual JWT token.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"